docs(examples): update single_threaded example to async/await
